verify bool return value (#4)
Boolean flag values are handled special in the current PostHog SDK
implementation.

When a feature flag is not found, the returned value will be `false`.
However, for boolean flag values, the returned value will also be a
boolean, either `true` or `false` depending on the flag resolution.
Since we cannot distinguish the case if the flag is found or not for
boolean values, the explicit error is removed from the boolean flag
value. This way we avoid too many false-negatives.